A few more bugfixes.
Main issue fixed is that deploy() now gets the schema right when mapping the output of SQL queries to its internal representation.
