+++ src/test/resources/logback-test.xml
@@ -14,16 +14,17 @@
   <logger name="mimir.algebra.OperatorUtils" level="INFO"/>
   <logger name="mimir.ctables.CTPercolator" level="INFO"/>
   <logger name="mimir.ctables.CTExplainer" level="INFO"/>
-  <logger name="mimir.exec.Compiler" level="INFO"/>
+  <logger name="mimir.exec.Compiler" level="WARN"/>
   <logger name="mimir.exec.BestGuesser" level="INFO"/>
-  <logger name="mimir.exec.TupleBundler" level="DEBUG"/>
+  <logger name="mimir.exec.TupleBundler" level="INFO"/>
   <logger name="mimir.exec.WorldBits" level="INFO"/>
   <logger name="mimir.exec.result.ProjectionResultIterator" level="INFO"/>
   <logger name="mimir.exec.result.AggregateResultIterator" level="INFO"/>
-  <logger name="mimir.exec.result.JDBCResultIterator" level="INFO"/>
+  <logger name="mimir.exec.result.JDBCResultIterator" level="WARN"/>
   <logger name="mimir.lenses.BestGuessCache" level="INFO"/>
   <logger name="mimir.lenses.ShredderLens" level="INFO"/>
   <logger name="mimir.lenses.TypeInferenceLens$" level="INFO"/>
+  <logger name="mimir.lenses.KeyRepairLens$" level="INFO"/>
   <logger name="mimir.models.ModelManager" level="INFO"/>
   <logger name="mimir.models.EditDistanceMatchModel" level="INFO"/>
   <logger name="mimir.models.TypeInferenceModel" level="INFO"/>
