+++ src/test/scala/mimir/models/TypeInferenceModelSpec.scala
@@ -24,6 +24,7 @@ object TypeInferenceModelSpec extends SQLTestSpecification("TypeInferenceTests")
   {
     model.bestGuess(0, List[PrimitiveValue](), List()) match {
       case TypePrimitive(t) => t
+      case x => throw new RAException(s"Type inference model guessed a non-type primitive: $x")
     }
   }
